+++ src/models/session.rs
@@ -36,11 +36,11 @@ pub struct Session {
 	)]
 	pub last_active_organization_id: Option<Option<String>>,
 	#[serde(rename = "last_active_at")]
-	pub last_active_at: i32,
+	pub last_active_at: i64,
 	#[serde(rename = "expire_at")]
-	pub expire_at: i32,
+	pub expire_at: i64,
 	#[serde(rename = "abandon_at")]
-	pub abandon_at: i32,
+	pub abandon_at: i64,
 	/// Unix timestamp of last update.
 	#[serde(rename = "updated_at")]
 	pub updated_at: i64,
@@ -56,9 +56,9 @@ impl Session {
 		user_id: String,
 		client_id: String,
 		status: Status,
-		last_active_at: i32,
-		expire_at: i32,
-		abandon_at: i32,
+		last_active_at: i64,
+		expire_at: i64,
+		abandon_at: i64,
 		updated_at: i64,
 		created_at: i64,
 	) -> Session {
